server/requirements.txt
fastapi
uvicorn
livekit-api
python-dotenv
aiohttp
pydantic
